src/test/java/ru/job4j/generics/question/AnalyzeTest.java
package ru.job4j.generics.question;

import org.junit.Test;

import java.util.Set;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.*;
import static ru.job4j.generics.question.Analyze.*;

public class AnalyzeTest {

    @Test
    public void whenNoDiff() {
        User u1 = new User(1, "A");
        User u2 = new User(2, "B");
        User u3 = new User(3, "C");
        Set<User> previous = Set.of(u1, u2, u3);
        Set<User> current = Set.of(u1, u2, u3);
        assertThat(diff(previous, current),
                is(new Info(0, 0, 0)
                )
        );
    }

    @Test
    public void whenOneChange() {
        User u1 = new User(1, "A");
        User u2 = new User(2, "B");
        User u3 = new User(3, "C");
        Set<User> previous = Set.of(u1, u2, u3);
        Set<User> current = Set.of(u1, new User(2, "BB"), u3);
        assertThat(diff(previous, current),
                is(new Info(0, 1, 0)
                )
        );
    }

    @Test
    public void whenOneAdded() {
        User u1 = new User(1, "A");
        User u2 = new User(2, "B");
        User u3 = new User(3, "C");
        Set<User> previous = Set.of(u1, u2, u3);
        Set<User> current = Set.of(u1, u2, u3, new User(4, "D"));
        assertThat(diff(previous, current),
                is(new Info(1, 0, 0)
                )
        );
    }

    @Test
    public void whenOneDeleted() {
        User u1 = new User(1, "A");
        User u2 = new User(2, "B");
        User u3 = new User(3, "C");
        Set<User> previous = Set.of(u1, u2, u3);
        Set<User> current = Set.of(u1, u2);
        assertThat(diff(previous, current),
                is(new Info(0, 0, 1)
                )
        );
    }
}